windows_distribution/source/requirements.txt
# Core dependencies with version pinning for stability
shotgun_api3==3.3.5
PyQt5==5.15.10
PyYAML==6.0.1
pillow==10.4.0
python-dotenv==1.0.1

# Optional dependencies
pyexiftool==0.5.6
ffmpeg-python==0.2.0
markdown==3.6

# Development/testing
pytest==8.3.2
